refactor(web): standardize hotkey ownership
