+++ web/app/components/snippets/components/snippet-header/run-mode.tsx
@@ -4,8 +4,8 @@ import { cn } from '@langgenius/dify-ui/cn'
 import * as React from 'react'
 import { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
-import { TEST_RUN_MENU_HOTKEY } from '@/app/components/workflow/header/shortcuts.ts'
 import { useWorkflowRun, useWorkflowStartRun } from '@/app/components/workflow/hooks'
+import { TEST_RUN_MENU_HOTKEY } from '@/app/components/workflow/hotkeys'
 import { ShortcutKbd } from '@/app/components/workflow/shortcuts/shortcut-kbd'
 import { useStore } from '@/app/components/workflow/store'
 import { WorkflowRunningStatus } from '@/app/components/workflow/types'
