--- web/app/components/workflow/header/shortcuts.ts
@@ -1,4 +1,0 @@
-import type { Hotkey } from '@tanstack/react-hotkeys'
-
-export const TEST_RUN_MENU_HOTKEY = 'Alt+R' satisfies Hotkey
-export const VERSION_HISTORY_HOTKEY = 'Mod+Shift+H' satisfies Hotkey
