+++ web/app/components/workflow/panel-contextmenu.tsx
@@ -8,7 +8,6 @@ import {
 import { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 import { FlowType } from '@/types/common'
-import { TEST_RUN_MENU_HOTKEY } from './header/shortcuts'
 import {
   useDSL,
   useIsChatMode,
@@ -17,6 +16,7 @@ import {
   useWorkflowStartRun,
 } from './hooks'
 import { useHooksStore } from './hooks-store'
+import { TEST_RUN_MENU_HOTKEY } from './hotkeys'
 import { isSnippetCanvas } from './nodes/_base/hooks/snippet-input-field-vars'
 import AddBlock from './operator/add-block'
 import { useOperator } from './operator/hooks'
